grammar: Rework thread-safe lexer/parser patching
The ANTLR generated code is not thread-safe. This was fixed by moving
some globals into the lexer/parser constructor by applying a patch to
the generated code. However, this is fragile as the patch often wont
apply after any non-trivial lexer/parser change.

Instead, provide a new constructor for the lexer and parser called
`NewThreadSafeSyslLexer()` and `NewThreadSafeSyslParser()` that create
the local data structures and override the global ones installed in the
lexer and parser. Call these new constructors where the old ones were
used.