grammar: Regenerate SyslParser.tokens
After the change in da6efd4 which added some `E_RAW_TEXT` tokens, the
`SyslParser.tokens` file was not committed. This is because the
`Makefile` did not copy it from the temp directory where it was
generated. Now the temp directory has gone, it is noticeable that the
file has changed.