grammar: Add sysl_parser.go dependency on SyslLexer.tokens
Generating the SyslParser ANTLR code requires that the lexer be
regenerated if there are any new tokens. The parser gets the tokens from
the `SyslLexer.tokens` file which is generated when the lexer code is
generated. Add a dependency to the parser on that lexer tokens file so
that it is regenerated if needed.