build: Set build go version/os from make build
`make build` was setting the build date, version and git commit, but not
the go build version and arch. These are set in the goreleaser config.
Set them in the `Makefile` too so local and docker builds set these
variables.

This makes `docker run sysl info` show the build version and os/arch.