build: Make Dockerfile actually build sysl
The sysl dockerfile was copying the `sysl` binary from outside the
docker environment, but adding it to an incompatible image. The build
environment would have been github actions or goreleaser, which is
glibc-based, but was added to an Alpine image, which is musl-based.

Instead of copying in a pre-built binary, use a multistage Dockerfile to
build in a compatible container image.

Clean up the `.dockerignore` so that we can actually build the source
code, as it was ignoring the `pkg/` directory and add a few other
entries to be ignored as they are generated files. Also ignore the docs
as they are not likely to be needed in a Dockerfile. Finally ignore a
couple of large files that are not needed (the antlr jar file and
`scripts/generate`). It is possible they'll be needed at some time, but
they can be removed then. This keeps the docker build context smaller,
making it faster to build.