parser: Add support for enumeration types
Add the `!enum` clause to the grammar to populate the Enum variant of
Type in the sysl protobuf ast.

The current syntax is provisional:

    app:
        !enum foo:
            bar: 1
            baz: 2

The use of a colon (`:`) to separate the name from the value was not the
first choice - that would be equals (`=`), but lines with text, numbers,
whitespace and EQ only are lexed as `TEXT_LINE` and cannot be parsed as
`Name EQ DIGITS` as desired. We expect we'll change this to `=` when the
new parser is ready.

An enum type can have attributes/modifiers. The individual values
cannot. The values of enums can only be integers (int64) and they must
have a value. This is due to the sysl protobuf ast. We may want to
change that to allow string enums, enums without values (auto
calculated, or names just for documentation purposes). It would also be
good if we could attach docstrings/comments to each enumeration. Also
maintaining order of enumerations may be important.

The current use case for this parsed enum support is to go as far as the
sysl protobuf ast, where a new out-of-tree (currently) code generator
will take that protobuf to generate required files that need enums. This
means there is currently no code generation or diagram generation or
whatever for enums in this repo.

Update the language guide to describe enums.

Issues: #243