alias parses type correctly and restricts to one type
`alias` rule is changed so that it allows only one type definition to be aliased as before it allows multiple types.

The previous rule also allows annotation and type to be in one line, this change requires it to be in different lines now.

The previous change to `reference` also made `alias.sysl` not parsed as `\n     sequence of Term` was parsed as a `reference` before the change to the `reference` rule. This change made it so that `sequence of Term` is parsed as a `collection_type`.