removing whitespaces at the start of reference which fixes multi line field
Previous `reference` rule allows whitespaces at the start, this breaks a lot of indentation rule which make `alias.sysl` parsed incorrectly (it fails now) and fields with multiple properties break as the `reference` rule would consume the indentation instead of the indetation rule.

Removing the whitespaces in `reference` rule allows indentation rule to handle the whitespaces which makes fields with multiple properties work.