allow raw parsing view
this is the one of the first step in transitioning towards arrai to replace sysl transform
the grammar will just consume the raw text of transforms without parsing it in more detail
the plan is to let the old parser (ANTLR) to handle sysl transforms.

the next step will be to remove transform entirely and move it to arrai.