handles lines with random spaces between indented statements (#735)
handles lines with random spaces between indented statements

current indentation rule does not anticipate lines with random spaces. For example (with underscores replacing spaces)

```
____statement1
___

____statement2
```

The newly added regex rule `/{(\n (?:[\_\t]+ \n)?)+}` will handle lines with random spaces.
It is also added to the global `.wrapRE` rule so that it does not ignore the random spaces as sometimes it is parsed as the `level` rule
which consumes spaces before a statement is necessary to determine the level of indentation of the statement.