handles inline comments and fix a comment parsing
The new `InlineIndentComment` allows a comment before the indented statements. This is to handle
comments after a colon as many indented statements begin with a colon. More comment rules are
also added to allow parsing comments in many places.

TEXT rules are also added so that it stops consuming before an inline comment.

Also added support on parsing CRLF `\r\n`.