modules: Search relative to a root
Add a `root` parameter to `mod.Fs` so that when a root is specified, it
can be a module.

When running `sysl cmd --root example.com/module foo.sysl`, the file
`foo.sysl` was not being found in the module `example.com/module` as the
`mod.Fs` did not know of the root. So it only looked for `foo.sysl`.

What did work was `sysl cmd example.com/module/foo.sysl` which appears
to be how all modules were being referenced.

Propagate the root to the `mod.Fs` so we can search relative to a module
root.