codegen: Look for grammar relative to transform root
When the `codegen` subcommand loads a grammar (from the `--grammar`
flag), look for it relative to `--root-transform` instead of `--root`.

The grammar is closely tied to the transform so the transform root is
the more correct place to look for it.

This allows the following invocation:

    sysl codegen --root-transform github.com/anz-bank/sysl-go/codegen \
      --grammar grammar/go.gen.g --transform transform/svc_types.sysl \
      ...

It also means the transform root can be substituted with a filesystem
path instead of a module allowing the use of unpublished transforms,
making development of existing transforms easier.