Fix map types conversion in syslwrapper
Fixes an issue in syslwrapper where maps weren't correctly imported. Maps are defined like the following example:

    !type InventoryResponse[json_map_key="item_id"]:
            item_id <: string:
                @json_tag = "item_id"
            quantity <: int:
                @json_tag = "quantity"

This is non-ideal, but due to limitations in the grammar, that's how they were implemented.

In this commit, we update the json_map_key to fill the Types.PrimaryKey field introduced for relation types in the simplified syslwrapper.Type struct.

In addition, Tuples with a json_map_key attribute will be converted into the "map" type.