fixes attribute scope issues in type
Initially higher attributes replaces attributes that are in lower scope.
This shouldn't happen.

For example:
```sysl
App:
    !type attr:
        @a = 1
        i <: string:
            @a = 2
```
The resulted sysl module would replace the `a` attribute under the
field `i` with the value 1 which is higher in the scope.

This commit ensures that if attr values in lower scopes can not
be replaced by values in the higher scope.