fixes multiline annotation merging issue (#403)
* fixes multiline annotation merging issue

Initially multiline annotations were collected line by line and stored directly into
the annotations map. This causes problems as it merges all multiline annotations
of the same annotations defined multiple times. This commit fixes that by storing
it in a temporary variable, assigns it based on the annotation precedence rule, and
resets the temporary variable.

* lint fix

* increase coverage attempt

* moved anno merge tests for coverage

* remove old anno_merge test