Fixes incorrect reconstruction for nested files (#407)
* Fixes incorrect reconstruction for nested files

Previously, reconstructing a Sysl file from a relational file
assumed produced a format that functioned for Sysl files in
existing directories but not files that required parent
directories to be created. Consider the following example
previous and new outputs:

{'foo/bar.sysl': '...'}       # previous output
{'foo': {'bar.sysl': '...'}}  # new output

The previous output only creates the files, not the necessary
parent directories to those files.

This change also addresses a potential non-deterministic output
for files containing applications that do not specify a line
source number. Previously, applications without a line source
number would resolve to the same value for the purpose of sorting,
resulting in a non-deterministic output. This change uses the
application name in instances where a line source number isn't
present.

* Updates generated files