Support locating sysl.pb via environment variable (#441)
If the `SYSL_PB_PATH` env var is set, `sysl.load()` will first attempt to locate `sysl.pb` at that path. If it's not present, `load` will fail immediately. If the env var is not set, existing behaviour is preserved.

In any case, `sysl.pb` is never fetched over the network if `load` is not called.