GH-463 Improve source context calculation (#479)
The SourceContexts set on most Sysl elements has always been incorrect. It relies on the lines and columns reported by the Antlr parser context, which deals with non-rule tokens in unexpected ways.

As part of this mini-overhaul of SourceContexts, I've taken the liberty of **making both lines and columns zero-based.** Previously both were 1-based, so the starting Location of the first character was `(line: 1, col: 1)`. This is what editors display, but it doesn't seem right for the underlying data structure, since SourceContext is an index into a document. [VS Code also uses zero-based indexes for both lines and columns](https://docs.microsoft.com/en-us/dotnet/api/microsoft.visualstudio.languageserver.protocol.position?view=visualstudiosdk-2019). Antlr has 1-based lines and 0-based columns, which is just perverse. In any case, this is a breaking change, but I'm unaware of any code that depends on actual values of SourceContext; it is primarily used relatively for sorting by position in file.

Changes proposed in this pull request:
- Convert SourceContext Locations to 0-based indexes.
- Fix various SourceContext Location values that were being incorrectly determined (most issues were with end locations)
- Add SourceContexts to Exprs
- Generate more golden test outputs with `make golden`
- Regenerate many many golden test outputs

Note that this PR does not totally fix all SourceContext locations, but it's a big improvement. Further improvements will have diminishing returns, and will be obsoleted by the eventual arr.ai-based paser.