importer (#558)
This commit replaces the existing importer for openapi3 with an arrai implementation that works with existing openapi3 importer tests. The old importer has been moved to `openapi3_legacy.go`. They are not removed as it is still being used by swagger importer and the swagger requires more features that the new importer doesn't implement yet.