Allow Sysl source on stdin for sysl commands (#578)
Sysl binary commands that accept Sysl module(s) as input currently require at least one path to be given as an argument.

This PR allows those commands to take no paths as input, and in that case attempt to read the source of a module from stdin. If present, compile that to a module and proceed as usual.

This is useful in ephemeral contexts where source is not written to disk (e.g. a new tab in an IDE). Ideally the same mechanism would accept marshalled sysl.Module protobuf messages to skip the compile step.

For example:
```sh
$ echo "App:   
>   ..." | sysl pb --mode=textpb 

apps:  {
 key:  "App"
[...]
}
```