Update and tidy LSP impl, add lsp CLI command (#603)
Currently, the Sysl LSP language server implementation is a standalone Go program that is built and distributed separately from the Sysl binary. The VS Code extension that uses it requires users to have the Go runtime installed so that the extension can `go get` the LSP server.

It would already be an improvement to distribute the Sysl LSP server as a binary instead of expecting it to be built from source on demand. However since the extension requires the Sysl binary to be present anyway, bundling the LSP server in with the binary removes the need for explicit LSP distribution altogether.

The `lsp` command (i.e. `sysl lsp`) functions as an LSP server over stdio: it will read a JSON-RPC message from stdin, and send responses and notifications over stdout (to be managed by the client). This is the same behaviour as the existing standalone Sysl LSP server, and adds a mere 400KB (<1%) to an unoptimised binary.

As such, the standalone LSP server is now deprecated. The VS Code extension will presently be updated to use the new `sysl lsp` command, and in a month or so the old `sysllsp` tool will be removed altogether.

More details of this change:
- Update the existing LSP framework code by copying the latest from https://github.com/golang/tools/tree/master/internal/lsp and tweaking appropriately
- Remove superfluous comments in the copied files
- Remove existing files relating to debugging/telemetry to simplify the codebase (will be re-added only when needed)
- Add additional no-op handlers for seemingly required client requests to avoid errors
- Fix `server_gen.go` generation; include in `make generate`