Fix error with module stdin on transform (#609)
If a module is provided via stdin, it doesn't have a corresponding path. sysl transform expect each module to have a path and fails if they don't.

This PR defaults the path to "stdin" in the case where there are more modules than paths (currently only possible with modules via stdin).