Accept paths for modules on stdin
The current solution fails if the source to stdin performs a relative import, since it doesn't know where the source is, and thus from where to resolve the import.

Instead of simply piping source to stdin, this PR changes the expected format of the input to an array of file objects, each with a path and content. Each item in the array will be treated as a MODULES arg (though if MODULES is provided it will take precedence and stdin is ignored).