Use standard loader when parsing files from stdin (#617)
Currently `source_context.file` for files parsed via stdin is incorrect. The file paths are technically correct, but they are not normalized relative to the Sysl root as expected.

It turns out that `loader.LoadSyslModule` does some material work over and above `parse.NewParser().Parse` which is necessary to meet these expectations. This PR uses a CopyOnWriteFs to overlay the files from stdin over the OS file system during the parse in order to use the existing `LoadSyslModule` logic.