In Go-based normalize, handle raw return statement payload with annos (#652)
When a return statement payload contains an invalid Sysl reference, the "raw" grammar rule parses it. However the "raw" rule doesn't handle [, so if a payload that must fall back on the "raw" rule has annotations as well, the "raw" rule will fail, as will the overall parse.

This change allows for attributes on ref and raw types (with the $ workaround).