bundles all transform script and move them to internal/bundles (#663)
Previously, all sysl transform scripts were used by reading the arr.ai files. This causes problems in arr.ai importing as arr.ai scripts can import from root, but it expects there to be a root marker (currently a go.mod file). When the root import is used and the sysl transform is called in a directory without an arr.ai root marker, sysl would fail.

This commit bundles all the scripts and moved it to internal/bundles. It also moves all the existing tests and redirected the filesystem to the one defined in internal/bundles/bundles.go.