ensure endpoint maps are defined in compilation (#670)
Previously, the parser panics when there is a subscribe endpoint to an application without endpoints. The application only had types and because of that, the endpoint maps were not defined. When parsing a subscribe endpoint, it then tries to assign to the nil endpoint map. This commit fixes that by ensuring that the endpoints maps are never nil.