Fix proto import resolution for local imports (#680)
When parsing import paths, the parser checks each subpath of the import path, including the bare `<name>.proto`, against the set of all proto paths available for import. Therefore an import of a "local" import (i.e. built into the parser) may be preferentially resolved as a file path import if _any_ subpath of the import matches an available proto on the file system.

For example, imagine one were to `import google/protobuf/empty.proto`, and also include `file/system/path/to/custom/empty.proto` in the available import paths. The parser would first check for matches for `google/protobuf/empty.proto`, `protobuf/empty.proto` and `empty.proto` on the file system, and in the last case it would match `file/system/path/to/custom/empty.proto` instead of `google/protobuf/empty.proto`. Thus `google/protobuf/empty.proto` is dropped from the set of available imports, and at resolution time `google/protobuf/empty.proto` cannot be resolved.

The parser's behaviour is suspicious, but this PR resolves the issue with only a small change: check for exact matches in `localImports` in preference to partial matches in the file system.

Also:
- A more helpful error message is also logged when the import resolution fails.
- Trim trailing whitespace.