Update prettier config and run (#756)
Prettier is a common code formatting tool that works across multiple languages, so it's a convenient choice to keep our code style consistent and avoid spurious deltas in PRs.

By default it can be a bit too aggressive in breaking long lines. It recommends a printWidth of 80 which targets 80-char width, but isn't overly militant about going over 80 chars. In practice, it does tend to break more than we would like, so bumping it up to 100 helps to strike a better balance.