Support loading a precompiled Sysl module via stdin (#749)
`sysl` commands can accept an input module in three ways: path to source file, path to precompiled model (.pb or .json), or via stdin with a JSON blob of the form `[{path: <path/to/file.sysl>, content: <sysl source>}, ...]` where `content` is parsed and compiled as if it were loaded from `path`. This covers the main use cases.

However sometimes you might have the bytes of a precompiled model that isn't available on disk, in which case `sysl` commands should accept that via stdin.

Rather than messing with the `stdinFiles` JSON blob format, this PR adds a fallback: if the content on stdin doesn't parse as a `stdinFiles` array, then try treating it as the bytes of a `.pb` model and unmarshalling it into a `*sysl.Module`.

This new mechanism is exposed to the TS lib as well via `PbDocumentModel.fromPb`, which passes a buffer of .pb file bytes to the `sysl` command's stdin.