Merge statements in endpoints when importing to existing files (#835)
Add TypeScript wrapper around Sysl CLI.

In particular, when invoking the `import` command, do some extra processing to merge imports and endpoints from the output file (if it already exists) and thus not clobber manual changes to those parts of the spec. This allows users to capture statements in imported endpoints.