change allOf to just resolve the type when there is only one type (#851)
This is a change in openapi behavior to allow some unhandled types in allOf. When there is only one type in allOf, resolve the type directly instead of trying to map it to an object. That way some previously unmapped type in allOf can be mapped but only when there is only one type.