Try harder to choose a response type name if it defaults to 'object' (#867)
When an anonymous endpoint response type is defined as having fields, we generate a type called object to represent it. If multiple endpoints do this, we generate multiple object types that get incorrect merged.

This change assigns a unique name to each anonymous response type, using the endpoint's operationId and status if available, or the base response type name otherwise. I'm not sure whether this covers everything, but it does the job I need now.